hide alarm-entry from the rest of netdata
