+++ src/health/health_notifications.c
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include "health_internals.h"
+#include "health-alert-entry.h"
 
 // the queue of executed alarm notifications that haven't been waited for yet
 static struct {
